+++ packages/ricos-editor/web/src/utils/renderBlockComponent.tsx
@@ -1,8 +1,6 @@
 import type { FC } from 'react';
 import React from 'react';
 
-const offset = { x: 40 };
-
 export const renderSideBlockComponent = (
   id: string,
   dir: string,
@@ -10,15 +8,15 @@ export const renderSideBlockComponent = (
 ) => {
   const node = document.querySelectorAll(`[data-offset-key="${id}-0-0"]`)[0] as HTMLElement;
   if (node?.offsetParent) {
-    const { top, x } = node.getBoundingClientRect();
-    const parentTop = node.offsetParent.getBoundingClientRect().top;
+    const { top } = node.getBoundingClientRect();
+    const { top: parentTop, width } = node.offsetParent.getBoundingClientRect();
+    const insetInlineStart = (dir === 'ltr' ? width : 0) + 10;
     return (
       <div
-        dir={dir}
         style={{
           position: 'absolute',
           top: top - parentTop,
-          insetInlineEnd: x - offset.x,
+          insetInlineStart,
         }}
       >
         <Component id={id} />